# according to benchmarks speed up by about 7-8% for pubkey computation