--- a/include/secp256k1_silentpayments.h
+++ b/include/secp256k1_silentpayments.h
@@ -7,7 +7,26 @@
 extern "C" {
 #endif
 
-/* TODO: add module description */
+/* This module provides an implementation for the ECC related parts of
+ * Silent Payments, as specified in BIP352. This particularly involves
+ * the creation of input tweak data by summing up private or public keys
+ * and the derivation of a shared secret using Elliptic Curve Diffie-Hellman.
+ * Combined are either:
+ *   - spender's private keys and receiver's public key (a * B, sender side)
+ *   - spender's public keys and receiver's private key (A * b, receiver side)
+ * With this result, the necessary key material for ultimately creating/scanning
+ * or spending Silent Payment outputs can be determined.
+ *
+ * Note that this module is _not_ a full implementation of BIP352, as it
+ * inherently doesn't deal with higher-level concepts like addresses, output
+ * script types or transactions. The intent is to provide cryptographical
+ * helpers for low-level calculations that are most error-prone to custom
+ * implementations (e.g. enforcing the right y-parity for key material, ECDH
+ * calculation etc.). For any wallet software already using libsecp256k1, this
+ * API should provide all the functions needed for a Silent Payments
+ * implementation without the need for any further manual elliptic-curve
+ * operations.
+ */
 
 /* TODO: add function API for sender side. */