--- a/src/modules/silentpayments/main_impl.h
+++ b/src/modules/silentpayments/main_impl.h
@@ -9,6 +9,37 @@
 #include "../../../include/secp256k1.h"
 #include "../../../include/secp256k1_silentpayments.h"
 
+/** Sort an array of silent payment recipients. This is used to group recipients by scan pubkey to
+ *  ensure the correct values of k are used when creating multiple outputs for a recipient. */
+static int secp256k1_silentpayments_recipient_cmp(const void* r1, const void* r2, void *cmp_data) {
+    const secp256k1_silentpayments_recipient *first = r1;
+    const secp256k1_silentpayments_recipient *second = r2;
+    secp256k1_ec_pubkey_sort_cmp_data *cmpd = cmp_data;
+    return secp256k1_ec_pubkey_cmp(cmpd->ctx, &first->scan_pubkey, &second->scan_pubkey);
+}
+
+int secp256k1_silentpayments_recipient_sort(const secp256k1_context* ctx, const secp256k1_silentpayments_recipient **recipients, size_t n_recipients) {
+    secp256k1_ec_pubkey_sort_cmp_data cmp_data;
+    VERIFY_CHECK(ctx != NULL);
+    ARG_CHECK(recipients != NULL);
+
+    cmp_data.ctx = ctx;
+
+    /* Suppress wrong warning (fixed in MSVC 19.33) */
+    #if defined(_MSC_VER) && (_MSC_VER < 1933)
+    #pragma warning(push)
+    #pragma warning(disable: 4090)
+    #endif
+
+    secp256k1_hsort(recipients, n_recipients, sizeof(recipients), secp256k1_silentpayments_recipient_cmp, &cmp_data);
+
+    #if defined(_MSC_VER) && (_MSC_VER < 1933)
+    #pragma warning(pop)
+    #endif
+
+    return 1;
+}
+
 /* TODO: implement functions for sender side. */
 
 /* TODO: implement functions for receiver side. */