… .cirrus.yml
… .github/workflows/ci.yml
… CMakeLists.txt
… src/hsort_impl.h
… src/modules/extrakeys/Makefile.am.include
… src/modules/silentpayments/tests_impl.h
… src/modules/silentpayments/vectors.h
… tools/bech32m.py
… tools/ripemd160.py
… tools/tests_silentpayments_generate.py