This change reduces the time needed to generate pubkeys from tweaks by ~9% (see benchmarks [here]).(https://github.com/setavenger/libsecp256k1-sp-example/blob/optimise_sp_module/bench_results/bench-optimised-module-100-iter.txt)

In concrete numbers from my machine:

The setup were 10 runs with 100 iterations each. One iteration generated pub keys from 861 tweaks (tweaks were taken from block 834761).

Optimised Implementation:
    Mean: 34.9741 seconds
    Median: 34.9358 seconds
    Standard Deviation: 0.4609 seconds

Base Implementation:
    Mean: 38.7348 seconds
    Median: 38.5146 seconds
    Standard Deviation: 0.9978 seconds